+++ cli/src/command/log/mod.rs
@@ -1,12 +1,16 @@
 use std::{
-  fmt::Write as _,
+  fmt::{Display, Write as _},
   io,
   sync::{Arc, Mutex, MutexGuard},
+  time::Duration,
 };
+
+use chrono::{DateTime, Local};
+use colored::{Color, Colorize as _};
 
 use is_terminal::IsTerminal;
 use lapse::{
-  log::ResponseLogsIter,
+  log::{ResponseLog, ResponseLogsIter},
   tree::{FlatTreeConfig, resource::Resource},
 };
 
@@ -19,17 +23,74 @@ use minus::{Pager, hooks::Hook};
 
 pub mod error;
 
+pub struct FormatedLogEntry(ResponseLog);
+
+impl FormatedLogEntry {
+  pub fn new(src: ResponseLog) -> Self {
+    Self(src)
+  }
+}
+
+fn status_color(status: u16) -> colored::Color {
+  match status {
+    200 | 201 | 204 => Color::Green,
+    500 | 503 | 400 | 404 | 403 | 401 => Color::Red,
+    _ => Color::Blue,
+  }
+}
+
+impl Display for FormatedLogEntry {
+  fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+    writeln!(f, "{} {}", "Request".black(), self.0.request.green())?;
+
+    let curr_time = Local::now();
+    let curr_timezone = curr_time.timezone();
+
+    let dt = DateTime::from_timestamp_nanos(self.0.timestamp as i64).with_timezone(&curr_timezone);
+    let formated_date = dt.format("%d-%m-%Y %H:%M:%S").to_string();
+    writeln!(f, "{} {}", "Date:".black(), formated_date.cyan())?;
+
+    let base_duration = Duration::from_nanos_u128(self.0.duration);
+    let duration_milis = base_duration.as_millis();
+    writeln!(
+      f,
+      "{} {}{}",
+      "Duration:".black(),
+      duration_milis.to_string().purple(),
+      "ms".purple()
+    )?;
+    writeln!(
+      f,
+      "{} {}\n",
+      "Status:".black(),
+      self.0.status.to_string().color(status_color(self.0.status))
+    )?;
+
+    writeln!(f, "{}\n", "Headers".red())?;
+
+    for (key, val) in &self.0.headers {
+      writeln!(f, "{}: {}", key.bright_black(), val.yellow())?;
+    }
+
+    writeln!(f)?;
+    writeln!(f, "{}", self.0.text)?;
+
+    Ok(())
+  }
+}
+
 fn append_to_log_until_fills(
   thread_pager: &mut Pager,
   mut missing_lines: usize,
   entries: &mut MutexGuard<'_, ResponseLogsIter>,
 ) -> crate::Result<()> {
   while missing_lines > 0 {
-    if let Some((entry_name, new_log)) = entries.next() {
-      let lines_written = new_log.lines().count() + 1;
+    if let Some(new_log) = entries.next() {
+      let formated = FormatedLogEntry::new(new_log);
+      let content = format!("{formated}");
+      let lines_written = content.lines().count() + 1;
       missing_lines -= lines_written;
-      writeln!(thread_pager, "{}", entry_name).map_err(LogError::SendToPager)?;
-      writeln!(thread_pager, "{}", new_log).map_err(LogError::SendToPager)?;
+      writeln!(thread_pager, "{}", content).map_err(LogError::SendToPager)?;
     } else {
       break;
     }
@@ -46,7 +107,7 @@ pub fn log(request: Option<String>) -> crate::Result<()> {
   let selected_request = select_tree_entry(&tree, request, flatlist_config)?;
 
   if io::stdout().is_terminal() {
-    let entries_iter = lapse.logs_iter(&selected_request)?;
+    let entries_iter = lapse.logs_iter(&selected_request).into_parsed();
 
     let shared_entries_iter = Arc::new(Mutex::new(entries_iter));
     let pager = Pager::new();
@@ -88,9 +149,9 @@ pub fn log(request: Option<String>) -> crate::Result<()> {
 
     minus::page_all(pager).map_err(LogError::SetupPagerConfig)?;
   } else {
-    let entries_iter = lapse.logs_iter(&selected_request)?;
-    for (name, entry) in entries_iter {
-      println!("{name}");
+    let entries_iter = lapse.logs_iter(&selected_request).into_parsed();
+    let pretty_entries = entries_iter.map(FormatedLogEntry);
+    for entry in pretty_entries {
       println!("{entry}");
     }
   }